ctrl: rtestate: push wait helper into rtestate
The rtestate knows if machineconfigs should be created,
updated or deleted, so it is more consistent if it also
tell how to wait for the machineconfigs to be updated.

Push the waiting function logic into rtestate, from controller.
No intended change in behavior.
