+++ pkg/objectstate/rte/rte.go
@@ -71,10 +71,12 @@ type ExistingManifests struct {
 	enableMachineConfig bool
 }
 
-func (em *ExistingManifests) MachineConfigsState(mf rtemanifests.Manifests) []objectstate.ObjectState {
+type MCPWaitForUpdatedFunc func(string, *machineconfigv1.MachineConfigPool) bool
+
+func (em *ExistingManifests) MachineConfigsState(mf rtemanifests.Manifests) ([]objectstate.ObjectState, MCPWaitForUpdatedFunc) {
 	var ret []objectstate.ObjectState
 	if mf.MachineConfig == nil {
-		return ret
+		return ret, nullMachineConfigPoolUpdated
 	}
 	for _, tree := range em.trees {
 		for _, mcp := range tree.MachineConfigPools {
@@ -122,7 +124,50 @@ func (em *ExistingManifests) MachineConfigsState(mf rtemanifests.Manifests) []ob
 		}
 	}
 
-	return ret
+	return ret, em.getWaitMCPUpdatedFunc()
+}
+
+func (em *ExistingManifests) getWaitMCPUpdatedFunc() MCPWaitForUpdatedFunc {
+	if em.enableMachineConfig {
+		return IsMachineConfigPoolUpdated
+	}
+	return IsMachineConfigPoolUpdatedAfterDeletion
+}
+
+func nullMachineConfigPoolUpdated(instanceName string, mcp *machineconfigv1.MachineConfigPool) bool {
+	return true
+}
+
+func IsMachineConfigPoolUpdated(instanceName string, mcp *machineconfigv1.MachineConfigPool) bool {
+	existing := isMachineConfigExists(instanceName, mcp)
+
+	// the Machine Config Pool still did not apply the machine config wait for one minute
+	if !existing || machineconfigv1.IsMachineConfigPoolConditionFalse(mcp.Status.Conditions, machineconfigv1.MachineConfigPoolUpdated) {
+		return false
+	}
+
+	return true
+}
+
+func IsMachineConfigPoolUpdatedAfterDeletion(instanceName string, mcp *machineconfigv1.MachineConfigPool) bool {
+	existing := isMachineConfigExists(instanceName, mcp)
+
+	// the Machine Config Pool still has the machine config return false
+	if existing || machineconfigv1.IsMachineConfigPoolConditionFalse(mcp.Status.Conditions, machineconfigv1.MachineConfigPoolUpdated) {
+		return false
+	}
+
+	return true
+}
+
+func isMachineConfigExists(instanceName string, mcp *machineconfigv1.MachineConfigPool) bool {
+	mcName := objectnames.GetMachineConfigName(instanceName, mcp.Name)
+	for _, s := range mcp.Status.Configuration.Source {
+		if s.Name == mcName {
+			return true
+		}
+	}
+	return false
 }
 
 // GetMachineConfigLabel returns machine config labels that should be used under the machine config pool
